move combination into own submodule
