+++ src/pykeen/datasets/ea/__init__.py
@@ -1,0 +1,1 @@
+"""Entity Alignment datasets."""
